+++ grails-app/domain/be/cytomine/command/secusersecrole/DeleteSecUserSecRoleCommand.groovy
@@ -8,22 +8,15 @@ import java.util.prefs.BackingStoreException
 import be.cytomine.security.User
 import be.cytomine.security.SecRole
 import be.cytomine.command.SimpleCommand
+import be.cytomine.Exception.WrongArgumentException
 
 class DeleteSecUserSecRoleCommand extends DeleteCommand implements SimpleCommand {
 
     def execute() {
-        try {
-            def postData = JSON.parse(postData)
-            User user = User.read(postData.user)
-            SecRole role = SecRole.read(postData.role)
-            SecUserSecRole userRole = SecUserSecRole.findBySecUserAndSecRole(user, role)
-            return super.deleteAndCreateDeleteMessage(user.id,userRole,[user.id,role.id] as Object[])
-        } catch(NullPointerException e) {
-            log.error(e)
-            return [data : [success : false, errors : e.getMessage()], status : 404]
-        } catch(BackingStoreException e) {
-            log.error(e)
-            return [data : [success : false, errors : e.getMessage()], status : 400]
-        }
+        User user = User.read(json.user)
+        SecRole role = SecRole.read(json.role)
+        SecUserSecRole userRole = SecUserSecRole.findBySecUserAndSecRole(user, role)
+        if (!userRole) throw new WrongArgumentException("UserRole $user/$role was not found!")
+        return super.deleteAndCreateDeleteMessage(user.id, userRole, [user.id, role.id] as Object[])
     }
 }
